move headings html into table.ejs
