+++ lib/templates/table.ejs
@@ -1,4 +1,13 @@
+<% const Utils = require('../utils'); -%>
+
 <table id="results-3048m" style="clear: both;" class="widget-wrapper">
-  <thead><%- headingsHtml %></thead>
-  <tbody class="tk-time-tracker"><%- dataHtml %></tbody>
+    <thead>
+        <tr>
+            <th>Date</th>
+            <% Object.keys(timeEntries.totals).forEach((k) => { %>
+            <th colspan="2" style="padding-right: 14px;"><%= Utils.formatProjectHeading(projects, leaveTypes, k) %></th>
+            <% }); %>
+        </tr>
+    </thead>
+    <tbody class="tk-time-tracker"><%- dataHtml %></tbody>
 </table>
